restrict pages

+++ base/views.py
@@ -10,6 +10,8 @@ from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 # from django.http import HttpResponse
 # Create your views here.
 
@@ -20,6 +22,9 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')
@@ -64,6 +69,7 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method=='POST':
@@ -74,9 +80,13 @@ def createRoom(request):
     context={'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user!=room.host:
+        return HttpResponse('you are not allowed here')
 
     if request.method=='POST':
         form= RoomForm(request.POST, instance=room)
@@ -86,9 +96,13 @@ def updateRoom(request, pk):
     context={'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room= Room.objects.get(id=pk)
+
+    if request.user!=room.host:
+        return HttpResponse('you are not allowed here')
+
     if request.method=='POST':
         room.delete()
         return redirect('home')
